+++ components/brave_ads/core/internal/privacy/tokens/unblinded_payment_tokens/unblinded_payment_token_util.h
@@ -29,7 +29,7 @@ bool UnblindedPaymentTokenExists(
 
 bool UnblindedPaymentTokensIsEmpty();
 
-int UnblindedPaymentTokenCount();
+size_t UnblindedPaymentTokenCount();
 
 }  // namespace brave_ads::privacy
