+++ components/brave_ads/core/internal/privacy/tokens/unblinded_tokens/unblinded_token_util.cc
@@ -77,7 +77,7 @@ bool UnblindedTokensIsEmpty() {
   return ConfirmationStateManager::GetInstance().GetUnblindedTokens().IsEmpty();
 }
 
-int UnblindedTokenCount() {
+size_t UnblindedTokenCount() {
   return ConfirmationStateManager::GetInstance().GetUnblindedTokens().Count();
 }
